+++ src/character/player/luigi_world.ts
@@ -1,0 +1,20 @@
+namespace Charjs {
+    export class LuigiWorld extends AbstractMarioWorld {
+        colors = [
+            '',        // 0
+            '#000000', // 1
+            '#ffffff', // 2
+            '#0b4f22', // 3
+            '#86581e', // 4
+            '#371977', // 5
+            '#3af67d', // 6
+            '#2abe5c', // 7
+            '#ffde73', // 8
+            '#dea539', // 9
+            '#f6c79a', // 10
+            '#f59035', // 11
+            '#966df5', // 12
+            '#6d3ef5'  // 13
+        ];
+    }
+}
